+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refine include guard macro name for `analysis.hpp`
 - Add `ExpWeightedIirFilter` class
 - Add `printRealMat()` and `printRealVec()` functions
+- Add `sin_polyApprox()` function
 
 ## [0.1.0] - 2021/7/22